+++ src/brick.cpp
@@ -31,7 +31,7 @@ Brick Brick::get_rotated(const Brick& brick, int quarters_rotation)
     Brick rotated_brick{brick};
     for (auto& pixel : rotated_brick.pixels)
     {
-        for (const auto& i : irange<size_t>(quarters_rotation % 4))
+        for (const auto& i : irange<int>(quarters_rotation % 4))
         {
             swap(pixel.coords.x, pixel.coords.y);
             pixel.coords.x *= -1;
